+++ src/main/java/com/example/farmfarm_refact/repository/OrderRepository.java
@@ -11,6 +11,6 @@ import org.springframework.web.bind.annotation.RequestParam;
 import java.util.List;
 import java.util.Optional;
 
-public interface OrderRepository extends CrudRepository<OrderEntity, Integer> {
+public interface OrderRepository extends CrudRepository<OrderEntity, Long> {
 
 }
